ensure TaskRespotiry is called with correct params

+++ backend/todolist/src/main/java/com/koglin/todolist/domain/models/TaskModel.java
@@ -1,21 +1,23 @@
 package com.koglin.todolist.domain.models;
 
+import java.util.Objects;
+
 public class TaskModel {
-    private int id;
+    private Long id;
     private String description;
     private boolean done;
 
-    public TaskModel(int id, String description, boolean done) {
+    public TaskModel(Long id, String description, boolean done) {
         this.id = id;
         this.description = description;
         this.done = done;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
